refactor(inspection-service): extract inspection-sharing sub-service
Introduce the inspection sub-service family: a shared module (types,
constants, pure helpers), a base class holding the injected deps, and the
first focused sub-service (agent view token sharing). The InspectionService
class becomes a facade that delegates the two sharing methods. Public API
(class, constructor signature, named exports) is unchanged — moved code is
byte-identical.
